Move components to separate files, update styles for most of the component similar to Questionnaire
